chore: moved sources to chain module
