+++ src/chain/mod.rs
@@ -4,6 +4,7 @@ use tracing::instrument;
 mod create;
 mod delete;
 mod list;
+mod sources;
 mod update;
 
 #[derive(Parser)]
